Fix warnings from idirs 1.0.
The 4 warnings were around deprecated keywords and implicit binding.

./Data/Bifoldable.idr:11:8:
Data.Bifoldable.toDual has a name which may be implicitly bound.
This is likely to lead to problems!

The 'instance' keyword is deprecated. Use 'implementation' (or omit it) instead.

./Data/Bifoldable.idr:73:10:Use of deprecated name Prelude.Monad.return
Please use `pure`, which is equivalent.

The 'class' keyword is deprecated. Use 'interface' instead.